requirements.txt
ollama>=0.2.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pypdf>=4.0.0
colorama>=0.4.6
fastapi>=0.115.0
uvicorn>=0.30.0
